skills/automotive-examples-code/references/adas-perception/requirements.txt
# ADAS Perception Pipeline Dependencies

# Deep Learning Frameworks
torch==2.1.0
torchvision==0.16.0
ultralytics==8.0.220  # YOLO v8

# Computer Vision
opencv-python==4.8.1.78
opencv-contrib-python==4.8.1.78
Pillow==10.1.0

# Point Cloud Processing
open3d==0.18.0
numpy==1.24.3
scipy==1.11.4

# Sensor Fusion & Tracking
filterpy==1.4.5  # Kalman filter
scikit-learn==1.3.2  # DBSCAN clustering

# CARLA Integration (optional)
carla==0.9.15; platform_system != "Darwin"

# Visualization
matplotlib==3.8.2
seaborn==0.13.0

# Configuration & Utilities
PyYAML==6.0.1
tqdm==4.66.1
click==8.1.7

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Performance
numba==0.58.1  # JIT compilation

# Protobuf for output format
protobuf==4.25.1

# Optional: TensorRT for optimization
# nvidia-tensorrt==8.6.1
